requirements.txt
attrs==20.3.0
bcrypt==3.2.0
certifi==2020.12.5
cffi==1.14.4
chardet==3.0.4
click==7.1.2
cryptography==3.3.1
ecdsa==0.14.1
fastapi==0.62.0
h11==0.11.0
idna==2.10
importlib-metadata==3.3.0
iniconfig==1.1.1
packaging==20.8
passlib==1.7.4
pluggy==0.13.1
py==1.10.0
pyasn1==0.4.8
pycparser==2.20
pydantic==1.7.3
PyJWT==2.0.1
pyparsing==2.4.7
pytest==6.2.0
python-jose==3.2.0
python-multipart==0.0.5
requests==2.25.0
rsa==4.6
six==1.15.0
starlette==0.13.6
toml==0.10.2
typing-extensions==3.7.4.3
urllib3==1.26.2
uvicorn==0.13.1
zipp==3.4.0
